\name{Average-package}
\alias{Average-package}
\alias{Average}
\docType{package}
\title{
\packageTitle{Average}
}
\description{
\packageDescription{Average}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Average}
\packageIndices{Average}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{Average}

Maintainer: \packageMaintainer{Average}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
