\name{garbage}
\alias{garbage}
\docType{data}
\title{dati di base}
\description{sono i dati di base}
\usage{data("garbage")}
\format{
  A data frame with 9 observations on the following 2 variables.
  \describe{
    \item{\code{anno}}{a numeric vector}
    \item{\code{kg}}{a numeric vector}
  }
}
\details{
sapessi
}
\source{insulaR}
\references{insulaR}
\examples{
data(garbage)
garbage[,1]
}
\keyword{datasets}
