\name{geomean}
\alias{geomean}

\title{media geometrica}
\description{fa la media geom}
\usage{
geomean(x)
}

\arguments{
  \item{x}{
}
}
\details{
sapessi
}
\value{
NO
}
\references{
insulaR 2017
}
\author{
Nicola P
}
\note{
NA
}

\seealso{
NULLA
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
return(prod(x)^(1/length(x)))
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
