\name{harmean}
\alias{harmean}

\title{media armonica}
\description{
fa la media armonica
}
\usage{
harmean(x)
}

\arguments{
  \item{x}{
}
}
\details{
media
}
\value{
nullo
}
\references{
insulR 2017
}
\author{
Nicola
}
\note{
NA
}

\seealso{
nulla
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
return(length(x)/sum(1/x))
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
