\name{average-package}
\alias{average-package}
\alias{average}
\docType{package}
\encoding{UTF-8}
\title{Medie speciali}
\description{
Libreria esemplificativa di accompagnamento dei post Costruire e pubblicare
una libreria per R pubblicati su \url{www.insular.it}.
}
\details{
\tabular{ll}{
Package: \tab average\cr
Type: \tab Package\cr
Version: \tab 0.0-0\cr
Date: \tab 2014-09-15\cr
License: GPL-3\cr
}
La libreria contiene le funzioni \code{harmean} per il calcolo della media
armonica, \code{geomean} per il calcolo della media geometrica e il dataset
\code{garbage}.
}
\author{Davide Massidda <admin@insular.it>}
\references{
Costruire e pubblicare una libreria per R - parte 1: i file sorgente
\url{http://www.insular.it/?p=2755}
}
\keyword{package}
\seealso{\code{\link[average]{harmean}}, \code{\link[average]{geomean}}}
